Replace ADXL335 with MXR9500.
This commit includes a small change in the omc eagle library to make
replacing the ADXL335 with the MXR9500 easier. Basically I've swapped
the X,Y,Z pins and placed them closer together.

Changes in the schematics include replacing the accelerometer and
removing all traces of the self test pin (MXR9500 doesn't have this kind
of functionality).