Add RC low-pass filter for MXR9500 accelerometer.
The ADXL335 had its outputs trimmed to an impedance of 32kOhm. Adding
capacitors to the output resulted in a first-order low-pass RC filter.
When I replaced the ADXL335 with the MXR9500 I forgot to add resistors
to the low-impedance output pins of the sensor.
I changed the MXR9500 symbol in our OMC eagle library to make the
schematics look a bit nicer. There's a bit more distance between the
sensor output pins now.