Replace OPA4350 (SO14) with AD8544 (TSSOP14).
The buffer op-amps in the sensor circuitry have been replaced by the
AD8544, which suits our requirements better. It is optimized for supply
voltages between 2.7V and 5.5V, features rail-to-rail input and output,
is unity gain stable, available in a space-saving TSSOP14 package and to
make things even better quite low-cost.