Remove three motor connectors to save space.
In order to place all the parts on a 50x50 mm2 board trade-offs must be
made. Most people join the bus lines coming from the motors at one point
and make a single connection to the control board. Therefore we do not
need need four connectors on the board. A single one will be sufficient.