Add SN65HVD230 CAN driver to lib and schematics.
TI's SN65HVD230 is pin-compatible with Microchip's MCP2551 CAN bus
driver. In contrast to the MCP2551, it operates from 3.3V.

This commit adds the device to the OMC Eagle library and conducts the
necessary changes to replace the MCP2551 in the OMCmini schematics. All
resistors at the output of the driver have been removed, except for the
120 Ohm termination resistor.
Additionally the Rs pin has been connected to the STM32, which will allow
switching between normal operation and the listen-only mode, where the
transmitter stage is disconnected from the bus lines. This allows
recovering from certain error conditions.