Redesign power supply section.
Creating our own switching regulator with the TPS5286 is almost
impossible due to the space constraints we have.
The new approach uses a Recom R785.0-0.5 switching regulator which
includes all the necessary parts inside a small package. A LM3940 linear
regulator is used for the 5V -> 3.3V conversion.
A combination of TVS diode and polyfuse protects the board from inverse
polarity, overvoltage and overcurrent conditions.