Relocate signals to other STM32 pins. Add PPM.
I relocated a few signals to separate the analog signals from digital
I/O lines like SPI as good as possible.
I added a PPM signal, connected it to the Timer3 external counter pin on
one end and to the DSL connector on the other.
Added jumpers to DSL and PC USART to choose between VCC and VDD.