Redesign of the MCU pin assignment.
Analog signals are now subsumed in their own bus. I have tried to
connect them to the MCU in such a way that the analog signal will all be
in a single corner of the TQFP package instead of distributed across all
four sides.
The PPM input is now connected to PA8 (pin 41) where the capture channel
1 of Timer1 is placed.
I added a jumper to Boot0 which allows the user to choose between normal
startup from Flash and startup from system memory. This should enable
the bootloader support on USART1.
Last but not least, a lot of cleanup in the MCU section has been done.