Replace jumpers and ASE oscillator in schematics.
The SMD jumpers and the ASE oscillator came from the open-bldc lib. I
think it's better not to mix the parts from both eagle libraries. If
someone wants to work on the OMC (or open-bldc, resp.) schematics/
layouts he expects to find all the parts he needs in the corresponding
library.
In a further step we will merge the most important parts that both
projects share in the pretty-eagle-lib library. As an intermediate step
I copied some parts from the open-bldc to the OMC library, though.